prephub/requirements.txt
django==3.1.7
geosky==1.0.9
unidecode==1.2.0
djangorestframework==3.12.2